+++ resources/compiledJS/combat/combat.js
@@ -57,17 +57,10 @@ function buffOrHeal(character, ability) {
     }
     if (((_a = ability.statusesUser) === null || _a === void 0 ? void 0 : _a.length) > 0) {
         ability.statusesUser.forEach((status) => {
-            if (!character.statusEffects.find((eff) => eff.id == status)) {
-                // @ts-ignore
-                character.statusEffects.push(new statEffect(Object.assign({}, statusEffects[status]), ability.statusModifiers));
-                if (character.id == player.id)
-                    character.statusEffects.find((eff) => eff.id == status).last.current -= 1;
-            }
-            else {
-                character.statusEffects.find((eff) => eff.id == status).last.current += statusEffects[status].last.total;
-            }
-            // @ts-ignore
-            statusEffects[status].last.current = statusEffects[status].last.total;
+            const _Effect = new statEffect(Object.assign({}, statusEffects[status]), ability.statusModifiers);
+            if (character.id === "player")
+                _Effect.last.current -= 1;
+            character.addEffect(_Effect);
             spawnFloatingText(character.cords, ability.line, "crimson", 36);
             let string = "";
             if (character.id == "player")
@@ -145,17 +138,7 @@ function regularAttack(attacker, target, ability, targetCords, isAoe = false) {
             const resist = target.getStatusResists()[_Effect.type];
             const resisted = resist + helper.random(9, -9) > ability.status_power + helper.random(18, -18);
             if (!resisted) {
-                let missing = true;
-                target.statusEffects.forEach((effect) => {
-                    if (effect.id == status) {
-                        effect.last.current += _Effect.last.total;
-                        missing = false;
-                        return;
-                    }
-                });
-                if (missing) {
-                    target.statusEffects.push(Object.assign({}, _Effect));
-                }
+                target.addEffect(_Effect);
             }
             else {
                 spawnFloatingText(target.cords, "RESISTED!", "grey", 36);
@@ -164,29 +147,25 @@ function regularAttack(attacker, target, ability, targetCords, isAoe = false) {
     }
     if (((_j = ability.statusesUser) === null || _j === void 0 ? void 0 : _j.length) > 0) {
         ability.statusesUser.forEach((status) => {
+            const _Effect = new statEffect(Object.assign({}, statusEffects[status]), ability.statusModifiers);
+            if (attacker.id === "player")
+                _Effect.last.current -= 1;
             if (!attacker.statusEffects.find((eff) => eff.id == status)) {
-                // @ts-ignore
-                attacker.statusEffects.push(new statEffect(Object.assign({}, statusEffects[status]), ability.statusModifiers));
-                if (attacker.id == player.id)
-                    attacker.statusEffects.find((eff) => eff.id == status).last.current -= 1;
+                attacker.addEffect(_Effect);
+                spawnFloatingText(attacker.cords, ability.line, "crimson", 36);
+                if (!isAoe) {
+                    let string = "";
+                    if (attacker.id == "player")
+                        string = lang[ability.id + "_action_desc_pl"];
+                    else
+                        string = lang[ability.id + "_action_desc"];
+                    if (attacker.isFoe)
+                        displayText(`<c>crimson<c>[ENEMY] <c>yellow<c>${lang[attacker.id + "_name"]} <c>white<c>${string}`);
+                    else
+                        displayText(`<c>cyan<c>[ACTION] <c>white<c>${string}`);
+                }
             }
-            else {
-                attacker.statusEffects.find((eff) => eff.id == status).last.current += statusEffects[status].last.total;
-            }
-            // @ts-ignore
-            statusEffects[status].last.current = statusEffects[status].last.total;
-            spawnFloatingText(attacker.cords, ability.line, "crimson", 36);
-            if (!isAoe) {
-                let string = "";
-                if (attacker.id == "player")
-                    string = lang[ability.id + "_action_desc_pl"];
-                else
-                    string = lang[ability.id + "_action_desc"];
-                if (attacker.isFoe)
-                    displayText(`<c>crimson<c>[ENEMY] <c>yellow<c>${lang[attacker.id + "_name"]} <c>white<c>${string}`);
-                else
-                    displayText(`<c>cyan<c>[ACTION] <c>white<c>${string}`);
-            }
+            ;
         });
     }
     if (target.isFoe) {