+++ resources/compiledJS/items/consumable.js
@@ -1,7 +1,7 @@
 "use strict";
 class Consumable extends Item {
     constructor(base, setPrice = 0) {
-        var _a, _b, _c, _d;
+        var _a, _b, _c, _d, _e;
         super(base);
         const baseItem = Object.assign({}, items[this.id]);
         this.status = baseItem.status;
@@ -11,9 +11,11 @@ class Consumable extends Item {
         this.usesTotal = (_a = baseItem.usesTotal) !== null && _a !== void 0 ? _a : 1;
         this.usesRemaining = (_b = base.usesRemaining) !== null && _b !== void 0 ? _b : 1;
         this.equippedSlot = (_c = base.equippedSlot) !== null && _c !== void 0 ? _c : -1;
+        this.statusesUser = (_d = baseItem.statusesUser) !== null && _d !== void 0 ? _d : [];
+        this.modifiers = getAbiStatusModifiers(dummy, "attack", "dazed");
         this.stats = {};
         this.commands = {};
-        this.name = (_d = lang[this.id + "_name"]) !== null && _d !== void 0 ? _d : baseItem.name;
+        this.name = (_e = lang[this.id + "_name"]) !== null && _e !== void 0 ? _e : baseItem.name;
         if (setPrice > 0)
             this.price = setPrice;
         if (setPrice > 0) {