+++ script/enemy/enemies.ts
@@ -64,10 +64,7 @@ const enemies = {
       new Ability(abilities.attack, dummy),
     ],
     retreatLimit: 0, // when enemy has this % hp left, it runs away from the player once.
-    statsPerLevel: {
-      str: 1,
-      vit: 1
-    },
+    levelingTemplate: "slime_default",
     loot: [
       { type: "gold", amount: [1, 5] }
     ]
@@ -130,10 +127,7 @@ const enemies = {
       new Ability(abilities.attack, dummy)
     ],
     retreatLimit: 0, // when enemy has this % hp left, it runs away from the player once.
-    statsPerLevel: {
-      str: 1,
-      vit: 1
-    },
+    levelingTemplate: "slime_default",
     loot: [
       { type: "gold", amount: [4, 11] }
     ]
@@ -196,10 +190,7 @@ const enemies = {
       new Ability(abilities.attack, dummy)
     ],
     retreatLimit: 0, // when enemy has this % hp left, it runs away from the player once.
-    statsPerLevel: {
-      str: 1,
-      vit: 1
-    },
+    levelingTemplate: "slime_default",
     loot: [
       { type: "gold", amount: [4, 11] }
     ]
@@ -265,11 +256,7 @@ const enemies = {
       { id: "fast" }
     ],
     retreatLimit: 0, // when enemy has this % hp left, it runs away from the player once.
-    statsPerLevel: {
-      str: 1,
-      dex: 1,
-      vit: 2
-    },
+    levelingTemplate: "beast_melee",
     loot: [
       { type: "weapon", amount: [1, 1], item: "chippedBlade", chance: 20 },
       { type: "gold", amount: [2, 11] }
@@ -337,11 +324,7 @@ const enemies = {
       { id: "fast" }
     ],
     retreatLimit: 0, // when enemy has this % hp left, it runs away from the player once.
-    statsPerLevel: {
-      str: 2,
-      dex: 1,
-      vit: 1
-    },
+    levelingTemplate: "beast_melee",
     loot: [
       { type: "weapon", amount: [1, 1], item: "chippedBlade", chance: 20 },
       { type: "gold", amount: [2, 11] }
@@ -410,11 +393,7 @@ const enemies = {
       { id: "fast" }
     ],
     retreatLimit: 0, // when enemy has this % hp left, it runs away from the player once.
-    statsPerLevel: {
-      str: 1,
-      dex: 2,
-      vit: 1
-    },
+    levelingTemplate: "beast_ranged",
     loot: [
       { type: "weapon", amount: [1, 1], item: "hiisiBow", chance: 25 },
       { type: "gold", amount: [2, 11] }
@@ -478,11 +457,7 @@ const enemies = {
       { id: "magically_impotent" },
     ],
     retreatLimit: 0, // when enemy has this % hp left, it runs away from the player once.
-    statsPerLevel: {
-      str: 2,
-      dex: 1,
-      vit: 1
-    },
+    levelingTemplate: "skeleton_melee",
     loot: [
       { type: "weapon", amount: [1, 1], item: "chippedBlade", chance: 20 },
       { type: "gold", amount: [4, 15] }
@@ -547,11 +522,7 @@ const enemies = {
       { id: "no_natural_regen" },
       { id: "magically_impotent" },
     ],
-    statsPerLevel: {
-      str: 1,
-      dex: 2,
-      vit: 1
-    },
+    levelingTemplate: "skeleton_ranged",
     retreatLimit: 0, // when enemy has this % hp left, it runs away from the player once.
     loot: [
       { type: "weapon", amount: [1, 1], item: "huntingBow", chance: 20 },
@@ -617,11 +588,7 @@ const enemies = {
     statModifiers: [
       { id: "magical_regen" },
     ],
-    statsPerLevel: {
-      dex: 1,
-      int: 2,
-      vit: 1
-    },
+    levelingTemplate: "skeleton_mage",
     retreatLimit: 0, // when enemy has this % hp left, it runs away from the player once.
     loot: [
       { type: "gold", amount: [4, 15] }
@@ -689,11 +656,7 @@ const enemies = {
       { id: "magical_regen" },
       { id: "weaker_natural_ability" }
     ],
-    statsPerLevel: {
-      dex: 2,
-      int: 2,
-      vit: 1
-    },
+    levelingTemplate: "skeleton_mage",
     retreatLimit: 0, // when enemy has this % hp left, it runs away from the player once.
     loot: [
       { type: "armor", amount: [1, 1], item: "crownOfWisdom", chance: 5 },
@@ -761,11 +724,7 @@ const enemies = {
       { id: "no_natural_regen" },
       { id: "weaker_natural_ability" }
     ],
-    statsPerLevel: {
-      str: 2,
-      vit: 2,
-      cun: 1
-    },
+    levelingTemplate: "berserker_melee",
     loot: [
       { type: "weapon", amount: [1, 1], item: "chippedAxe", chance: 15 },
       { type: "weapon", amount: [1, 1], item: "stick", chance: 20 },
@@ -829,11 +788,7 @@ const enemies = {
     statModifiers: [
       { id: "no_natural_regen" },
     ],
-    statsPerLevel: {
-      dex: 2,
-      vit: 1,
-      cun: 2
-    },
+    levelingTemplate: "berserker_ranged",
     shootsProjectile: "hunterJavelinProjectile",
     loot: [
       { type: "gold", amount: [6, 18] }
@@ -902,11 +857,7 @@ const enemies = {
       { id: "orc_frenzy", },
       { id: "orc_resilience", }
     ],
-    statsPerLevel: {
-      str: 2,
-      vit: 2,
-      cun: 1
-    },
+    levelingTemplate: "berserker_melee",
     loot: [
       { type: "weapon", amount: [1, 1], item: "orcishAxe", chance: 15 },
       { type: "armor", amount: [1, 1], item: "ironShield", chance: 10 },
@@ -976,11 +927,7 @@ const enemies = {
       { id: "orc_frenzy", },
       { id: "orc_resilience", }
     ],
-    statsPerLevel: {
-      str: 2,
-      vit: 2,
-      cun: 1
-    },
+    levelingTemplate: "bulwark",
     loot: [
       { type: "weapon", amount: [1, 1], item: "chippedBlade", chance: 15 },
       { type: "armor", amount: [1, 1], item: "silverShield", chance: 8 },
@@ -1051,11 +998,7 @@ const enemies = {
       { id: "orc_frenzy", },
       { id: "orc_resilience", }
     ],
-    statsPerLevel: {
-      str: 3,
-      vit: 3,
-      cun: 1
-    },
+    levelingTemplate: "berserker_melee",
     loot: [
       { type: "weapon", amount: [1, 1], item: "galadorSpear", chance: 100 },
       { type: "gold", amount: [180, 260] }
@@ -1121,11 +1064,7 @@ const enemies = {
       { id: "magically_impotent" },
     ],
     retreatLimit: 0, // when enemy has this % hp left, it runs away from the player once.
-    statsPerLevel: {
-      str: 3,
-      dex: 0,
-      vit: 3
-    },
+    levelingTemplate: "beast_melee",
     loot: [
       { type: "gold", amount: [22, 44] }
     ]
@@ -1190,11 +1129,7 @@ const enemies = {
       { id: "magically_impotent" },
     ],
     retreatLimit: 0, // when enemy has this % hp left, it runs away from the player once.
-    statsPerLevel: {
-      str: 3,
-      dex: 0,
-      vit: 4
-    },
+    levelingTemplate: "beast_melee",
     loot: [
       { type: "gold", amount: [22, 44] }
     ]
@@ -1258,11 +1193,7 @@ const enemies = {
       { id: "magically_impotent" },
     ],
     retreatLimit: 0, // when enemy has this % hp left, it runs away from the player once.
-    statsPerLevel: {
-      str: 4,
-      dex: 0,
-      vit: 3
-    },
+    levelingTemplate: "beast_melee",
     loot: [
       { type: "weapon", amount: [1, 1], item: "trollClub", chance: 15 },
       { type: "gold", amount: [26, 51] }
@@ -1327,11 +1258,7 @@ const enemies = {
       { id: "magically_impotent" },
     ],
     retreatLimit: 0, // when enemy has this % hp left, it runs away from the player once.
-    statsPerLevel: {
-      str: 4,
-      dex: 0,
-      vit: 4
-    },
+    levelingTemplate: "beast_melee",
     loot: [
       { type: "weapon", amount: [1, 1], item: "trollClub", chance: 15 },
       { type: "gold", amount: [53, 129] }
@@ -1398,10 +1325,7 @@ const enemies = {
       { id: "slow" }
     ],
     retreatLimit: 0, // when enemy has this % hp left, it runs away from the player once.
-    statsPerLevel: {
-      str: 4,
-      vit: 4
-    },
+    levelingTemplate: "bulwark",
     loot: [
       { type: "gold", amount: [65, 150] }
     ]
@@ -1467,10 +1391,7 @@ const enemies = {
       { id: "slow" }
     ],
     retreatLimit: 0, // when enemy has this % hp left, it runs away from the player once.
-    statsPerLevel: {
-      str: 4,
-      vit: 4
-    },
+    levelingTemplate: "bulwark",
     loot: [
       { type: "weapon", amount: [1, 1], item: "stoneHalberd", chance: 12 },
       { type: "gold", amount: [90, 210] }
@@ -1537,10 +1458,7 @@ const enemies = {
       { id: "magically_impotent" },
     ],
     retreatLimit: 0, // when enemy has this % hp left, it runs away from the player once.
-    statsPerLevel: {
-      str: 2,
-      vit: 2
-    },
+    levelingTemplate: "balanced",
     loot: [
       { type: "weapon", amount: [1, 1], item: "silverSword", chance: 10 },
       { type: "offhand", amount: [1, 1], item: "silverShield", chance: 10 },
@@ -1614,11 +1532,7 @@ const enemies = {
       { id: "magically_impotent" },
     ],
     retreatLimit: 0, // when enemy has this % hp left, it runs away from the player once.
-    statsPerLevel: {
-      str: 2,
-      vit: 1,
-      cun: 1
-    },
+    levelingTemplate: "magical_melee",
     loot: [
       { type: "weapon", amount: [1, 1], item: "silverSword", chance: 10 },
       { type: "offhand", amount: [1, 1], item: "silverShield", chance: 10 },
@@ -1694,10 +1608,7 @@ const enemies = {
       { id: "weaker_natural_ability" }
     ],
     retreatLimit: 0, // when enemy has this % hp left, it runs away from the player once.
-    statsPerLevel: {
-      str: 3,
-      vit: 1
-    },
+    levelingTemplate: "magical_melee",
     loot: [
       { type: "gold", amount: [288, 541] }
     ]