+++ script/ui/icons.ts
@@ -166,4 +166,5 @@ const icons = {
   warning_icon: "resources/icons/warn.png",
   movementSpeed_icon: "resources/icons/speed.png",
   attackSpeed_icon: "resources/icons/speed.png",
+  liquid_courage_icon: "resources/icons/drunk.png",
 } as icons;