+++ Changelog.md
@@ -3,10 +3,12 @@
 ### Major update
 
 ## New Additions
+- Added 1 new consumable
 
 ## Fixes & Tweaks
 
 - Completely refactored code base to make working with it easier.
+- Status effects no longer add more time when gained, instead resetting the timer.
 
 # 1.0.9 ~ 16.3.2022